% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{medschool}
\alias{medschool}
\title{medschool}
\usage{
medschool()
}
\description{
medschool returns a family friendly Zoidberg quote.
}
\section{Warning}{

Do not operate heavy machinery within 8 hours of using this function.
}

\examples{
medschool()

}
