% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{nixon}
\alias{nixon}
\title{nixon}
\usage{
nixon()
}
\description{
nixon returns a quote from the head of ex-president Richard Nixon.
}
\section{Warning}{

The head-in-a-jar idea, at least in 2019, is only a philosophical thought
experiment. Please do not attempt.
}

