% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{squirrel}
\alias{squirrel}
\title{squirrel}
\usage{
squirrel()
}
\description{
squirrel returs a quote from the head of ex-president Richard Nixon.
}
\section{Note}{

No animals were harmed in the making of this quote.
Literally.
The squirrel did not fall, dispite Nixon's best efforts.
}

