% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{tempc2f}
\alias{tempc2f}
\title{Celsius to Fahrenheit}
\usage{
tempc2f(temp_c)
}
\arguments{
\item{temp_c}{}
}
\description{
Converts any temperature in Celcius to the
Fahrenheit equivalent. Because, well, life skills.
}
