% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{tempf2c}
\alias{tempf2c}
\title{Fahrenheit to Celsius}
\usage{
tempf2c(temp_F)
}
\arguments{
\item{temp_F}{}
}
\description{
Converts any temperature in Fahrenheit to the
Celcius equivalent. Because, well, life skills.
}
